function undefinePrimaryNetwork, to avoid the Failed to destroy network error added a check to see if the primary network is running

+++ functions.sh
@@ -283,7 +283,9 @@ function definePrimaryNetwork {
 }
 
 function undefinePrimaryNetwork {
-  virsh net-destroy primary
+  if primaryNetworkRunning;then
+    virsh net-destroy primary
+  fi
   virsh net-undefine primary
 }
